requirements.txt
admin-multiupload==1.10
asn1crypto==0.22.0
bcrypt==3.1.3
cffi==1.10.0
cryptography==2.0
Django==1.11.3
Fabric==1.10.2
idna==2.5
mysqlclient==1.3.10
paramiko==2.2.1
pep8==1.7.0
pyasn1==0.2.3
pycparser==2.18
PyNaCl==1.1.2
pytz==2017.2
six==1.10.0
